Extract getISODate() fn to dateTimeUtils file
